Move siteKitRequest to index.private.
